+++ changelog.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - fixed kernel.QueryDosDevice to support `None` as input parameter (returning
   the list of all MS-DOS device names)
+- checking the return code of kernel.GlobalUnlock
 
 ## [2.2.0] - 2023-05-27